requirements.txt
Django>=5.2,<6.0
pillow
pillow-heif
